i12e: 'butane/aux.go' split into 'butane/b64gzip.go' and 'butane/tplnew.go' and deleted
